fix warning by adding type casting (int) for %d

+++ lab2/sanf_data1.c
@@ -26,32 +26,32 @@ int main()
 	
 	// un-comment these lines if wish to run the code
 	
-	printf("size of c = %d, %d\n", sizeof(char), sizeof(c));
-	printf("size of cp = %d, %d\n", sizeof(*cp), sizeof(cp));
+	printf("size of c = %d, %d\n", (int) sizeof(char), (int) sizeof(c));
+	printf("size of cp = %d, %d\n", (int) sizeof(*cp), (int) sizeof(cp));
 
-	printf("size of s = %d, %d\n", sizeof(short), sizeof(s));
-	printf("size of sp = %d, %d\n", sizeof(*sp), sizeof(sp));
+	printf("size of s = %d, %d\n", (int) sizeof(short), (int) sizeof(s));
+	printf("size of sp = %d, %d\n", (int) sizeof(*sp), (int) sizeof(sp));
 
-	printf("size of i = %d, %d\n", sizeof(int), sizeof(i));
-	printf("size of ip = %d, %d\n", sizeof(*ip), sizeof(ip));
+	printf("size of i = %d, %d\n", (int) sizeof(int), (int) sizeof(i));
+	printf("size of ip = %d, %d\n", (int) sizeof(*ip), (int) sizeof(ip));
 
-	printf("size of l = %d, %d\n", sizeof(long), sizeof(l));
-	printf("size of lp = %d, %d\n", sizeof(*lp), sizeof(lp));
+	printf("size of l = %d, %d\n", (int) sizeof(long), (int) sizeof(l));
+	printf("size of lp = %d, %d\n", (int) sizeof(*lp), (int) sizeof(lp));
 
-	printf("size of ll = %d, %d\n", sizeof(long long), sizeof(ll));
-	printf("size of llp = %d, %d\n", sizeof(*llp), sizeof(llp));
+	printf("size of ll = %d, %d\n", (int) sizeof(long long), (int) sizeof(ll));
+	printf("size of llp = %d, %d\n", (int) sizeof(*llp), (int) sizeof(llp));
 	
-	printf("size of f = %d, %d\n", sizeof(float), sizeof(f));
-	printf("size of fp = %d, %d\n", sizeof(*fp), sizeof(fp));
+	printf("size of f = %d, %d\n", (int) sizeof(float), (int) sizeof(f));
+	printf("size of fp = %d, %d\n", (int) sizeof(*fp), (int) sizeof(fp));
 
-	printf("size of df = %d, %d\n", sizeof(double), sizeof(df));
-	printf("size of dfp = %d, %d\n", sizeof(*dfp), sizeof(dfp));
+	printf("size of df = %d, %d\n", (int) sizeof(double), (int) sizeof(df));
+	printf("size of dfp = %d, %d\n", (int) sizeof(*dfp), (int) sizeof(dfp));
 
-	printf("size of 100 = %d\n", sizeof(100));
+	printf("size of 100 = %d\n", (int) sizeof(100));
 
-	printf("size of 0xFFFFFFFF = %d\n", sizeof(0xFFFFFFFF));
+	printf("size of 0xFFFFFFFF = %d\n", (int) sizeof(0xFFFFFFFF));
 
-	printf("size of 0xFFFFFFFFFF = %d\n", sizeof(0xFFFFFFFFFF));
+	printf("size of 0xFFFFFFFFFF = %d\n", (int) sizeof(0xFFFFFFFFFF));
 
 //	WARNING: These code when un-commented will invoke compiler error:
 //	printf("size of v = %d, %d\n", sizeof(void), sizeof(v));
